requirements.txt
flask
openai
boto3
python-dotenv
fpdf
zappa
